gotw-backend/requirements.txt
fastapi
uvicorn
supabase
groq
pydantic
sentence-transformers
numpy
scikit-learn
python-dotenv
